增加了 " 角色数量，角色比例 " 的计算
测试无误

+++ src/main/scala/com/cw/flink/scala/pojo/PvUvHourModelToMySql.scala
@@ -3,14 +3,14 @@ package com.cw.flink.scala.pojo
 case class PvUvHourModelToMySql(
                            pv: Int,
                      uv:Int,
-                     manager_num_rate:Long,
-                     store_manager_num_rate:Long,
-                     supervisor_num_rate:Long,
+                     manager_num_rate:Double,
+                     store_manager_num_rate:Double,
+                     supervisor_num_rate:Double,
                      application_id: Int,
                      manager_num: Int,
                      store_manager_num: Int,
                      supervisor_num: Int,
-                     avg_stay_time: Long,
+                     avg_stay_time: Double,
                      hour: Int,
                      date_time: String,
                      create_time: String,
@@ -19,14 +19,14 @@ case class PvUvHourModelToMySql(
   override def productElement(n: Int): Any = n match {
     case 0 =>pv: Int
     case 1 => uv:Int
-    case 2 => manager_num_rate:Long
-    case 3 => store_manager_num_rate:Long
-    case 4 => supervisor_num_rate:Long
+    case 2 => manager_num_rate:Double
+    case 3 => store_manager_num_rate:Double
+    case 4 => supervisor_num_rate:Double
     case 5 => application_id: Int
     case 6 => manager_num: Int
     case 7 => store_manager_num: Int
     case 8 => supervisor_num: Int
-    case 9 => avg_stay_time: Long
+    case 9 => avg_stay_time: Double
     case 10 => hour: Int
     case 11 => date_time: String
     case 12 => create_time: String
